DM - Add Route Admin Get Product
Se añade ruta Get Product y se corrigen errores

+++ API/src/controllers/Products.js
@@ -3,6 +3,8 @@ const { upload, cloudinary } = require("../libs/storage");
 const { arrayProductos, arraycategorias } = require("./Data");
 
 const { Router } = require("express");
+const paginate = require("./Paginate");
+const { Op } = require("sequelize");
 const router = Router();
 
 const DB = async () => {
@@ -60,14 +62,15 @@ router.post("/", upload.array("image"), async (req, res, next) => {
         .send(
           "Formato de datos invalido (description) debe ser una cadena de texto."
         );
-    // console.log(image)
-    let imagenes = image.map((i) => i.path);
+     console.log(image)
+     
+    let imagenes = image ? image.map((i) => i.path) : undefined
     let product = await Product.create({
       name: name.toLowerCase(),
       description: description,
       price: price,
       stock: stock,
-      image: imagenes.length
+      image: imagenes
         ? imagenes
         : [
           "https://res.cloudinary.com/jdmoreno/image/upload/v1657596154/AppVinos/Default_hi3ylt.png",
@@ -100,8 +103,7 @@ router.post("/", upload.array("image"), async (req, res, next) => {
 router.put("/", upload.array("image"), async (req, res, next) => {
   try {
     const image = req.files || req.file;
-    const { id, name, description, price, stock, category } = req.body;
-
+    const { id, name, description, price, stock, category, state } = req.body;
     // if (!image) return res.status(400).send("Faltan datos necesarios (image).");
     if (!id) return res.status(400).send("Faltan datos necesarios (id).");
     if (!name) return res.status(400).send("Faltan datos necesarios (name).");
@@ -109,6 +111,7 @@ router.put("/", upload.array("image"), async (req, res, next) => {
       return res.status(400).send("Faltan datos necesarios (description).");
     if (!price) return res.status(400).send("Faltan datos necesarios (price).");
     if (!stock) return res.status(400).send("Faltan datos necesarios (stock).");
+    if (!state) return res.status(400).send("Faltan datos necesarios (state).");
     if (!category)
       return res.status(400).send("Faltan datos necesarios (category).");
     if (isNaN(parseInt(id)))
@@ -127,6 +130,10 @@ router.put("/", upload.array("image"), async (req, res, next) => {
       return res
         .status(400)
         .send("Formato de datos invalido (name) debe ser una cadena texto.");
+    if (!isNaN(parseInt(state)))
+      return res
+        .status(400)
+        .send("Formato de datos invalido (state) debe ser una cadena texto.");
     if (!isNaN(parseInt(description)))
       return res
         .status(400)
@@ -134,40 +141,42 @@ router.put("/", upload.array("image"), async (req, res, next) => {
           "Formato de datos invalido (description) debe ser una cadena de texto."
         );
 
-    let imagenes = image.map((i) => i.path);
     const prod = await Product.findByPk(parseInt(id));
     await prod.setCategories(category);
-    prod.image.map(async (i) => {
-      let m = i.split("/")[8].split(".")[0];
-      console.log(m);
-      await cloudinary.api.delete_resources(
-        "AppVinos/" + m,
-        function (error, result) {
-          console.log(result, error);
-        }
-      );
-    });
-    if (imagenes.length) {
+    if (image) {
+      let imagenes = image.map((i) => i.path);
+      prod.image.map(async (i) => {
+        let m = i.split("/")[8].split(".")[0];
+        console.log(m);
+        await cloudinary.api.delete_resources(
+          "AppVinos/" + m,
+          function (error, result) {
+            console.log(result, error);
+          }
+        );
+      });
       const product = await Product.update(
-        { name, description, price, stock, image: imagenes },
+        { name, description, price, stock, image: imagenes, state: state },
         {
           where: {
             id: id,
           },
         }
       );
-      if (product) return res.status(200).send("Producto actualizado.");
+      const prod = await Product.findByPk(id);
+      if (product) return res.status(200).json(prod);
       return res.status(400).send("Error al actualizar el producto.");
     } else {
       const product = await Product.update(
-        { name, description, price, stock },
+        { name, description, price, stock, state: state },
         {
           where: {
             id: id,
           },
         }
       );
-      if (product) return res.status(200).send("Producto actualizado.");
+      const prod = await Product.findByPk(id);
+      if (product) return res.status(200).json(prod);
       return res.status(400).send("Error al actualizar el producto.");
     }
   } catch (error) {
@@ -211,7 +220,6 @@ router.get("/carga", async (req, res, next) => {
 });
 
 router.get("/:id", async (req, res) => {
-  console.log('Carga')
   try {
     const { id } = req.params;
     if (id) {
@@ -227,6 +235,7 @@ router.get("/:id", async (req, res) => {
         price: productNew[0].price,
         stock: productNew[0].stock,
         image: productNew[0].image,
+        state: productNew[0].state,
         review: reviewNew,
         category: categoryNew
       }
@@ -237,6 +246,61 @@ router.get("/:id", async (req, res) => {
     res.status(400).send({ message: "Error: " + e })
   }
 });
+
+router.get("/", async (req, res) => {
+  try {
+    const { name, page, order_direction, order_by, category, page_limit } = req.query;
+    let search = {
+      where: {
+        state: "Activo"
+      }
+    }
+    let order = [];
+    let filter = {};
+    if (name) {
+      if (!isNaN(parseInt(name))) return res.status(400).json({ message: "Formato de datos invalido (name) debe ser una cadena texto." });
+      search = {
+        where: {
+          name: {
+            [Op.like]: `%${name.toLowerCase()}%`
+          },
+          state: "Activo"
+        }
+      }
+    }
+    if (category) {
+      if (isNaN(parseInt(category))) return res.status(400).json({ message: "Formato de datos invalido (category) debe ser un numero." });
+      filter = category;
+    }
+    if (order_direction) {
+      if (order_direction === 'DESC' || order_direction === 'ASC') {
+        order.push([['name', order_direction]])
+        if (order_by) {
+          if (order_by.toLowerCase() === 'price') {
+            order.pop();
+            order.push([['price', order_direction]])
+          }
+        }
+      } else {
+        return res.status(400).json({ message: "Datos invalidos (order_direction) permitidos: DESC o ASC." });
+      }
+    }
+
+    let data = await paginate(Product, page, page_limit, search, order, filter);
+    if (category) {
+      let data2 = []
+      await Promise.all(data.data.map(async (p) => {
+        data2.push(await getCategory(p.id));
+      }))
+      data = { ...data, data: data2 };
+    }
+
+    data.data.length ? res.status(200).json(data) : res.status(400).send({ message: "El vino descrito no se encuentra guardado" });
+
+  } catch (e) {
+    res.status(400).send({ message: "Error: " + e })
+  }
+})
 
 
 module.exports = router;
